# backend/openapi.json
{
  "components": {
    "schemas": {
      "AreaIstatOutput": {
        "properties": {
          "messaggio": {
            "title": "Messaggio",
            "type": "string"
          },
          "motivo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Motivo"
              },
              {
                "type": "null"
              }
            ]
          },
          "parametri": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParametriIstatOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "promemoria_manuale": {
            "title": "Promemoria Manuale",
            "type": "boolean"
          },
          "stato": {
            "$ref": "#/components/schemas/StatoConfigurazione"
          }
        },
        "required": [
          "stato",
          "motivo",
          "messaggio",
          "promemoria_manuale",
          "parametri"
        ],
        "title": "AreaIstatOutput",
        "type": "object"
      },
      "AreaTassaOutput": {
        "properties": {
          "messaggio": {
            "title": "Messaggio",
            "type": "string"
          },
          "motivo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Motivo"
              },
              {
                "type": "null"
              }
            ]
          },
          "parametri": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParametriTassaOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "promemoria_manuale": {
            "title": "Promemoria Manuale",
            "type": "boolean"
          },
          "stato": {
            "$ref": "#/components/schemas/StatoConfigurazione"
          }
        },
        "required": [
          "stato",
          "motivo",
          "messaggio",
          "promemoria_manuale",
          "parametri"
        ],
        "title": "AreaTassaOutput",
        "type": "object"
      },
...
      "ConfigurazioneNormativaOutput": {
        "properties": {
          "alla_data": {
            "format": "date",
            "title": "Alla Data",
            "type": "string"
          },
          "istat": {
            "$ref": "#/components/schemas/AreaIstatOutput"
          },
          "tassa_soggiorno": {
            "$ref": "#/components/schemas/AreaTassaOutput"
          }
        },
        "required": [
          "alla_data",
          "tassa_soggiorno",
          "istat"
        ],
        "title": "ConfigurazioneNormativaOutput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Motivo": {
        "enum": [
          "comune_non_riconosciuto",
          "comune_non_configurato",
          "regione_non_riconosciuta",
          "regione_non_configurata"
        ],
        "title": "Motivo",
        "type": "string"
      },
...
      "ParametriIstatOutput": {
        "properties": {
          "periodicita": {
            "$ref": "#/components/schemas/Periodicita"
          },
          "tracciato": {
            "title": "Tracciato",
            "type": "string"
          }
        },
        "required": [
          "tracciato",
          "periodicita"
        ],
        "title": "ParametriIstatOutput",
        "type": "object"
      },
      "ParametriTassaOutput": {
        "properties": {
          "esenzione_eta_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Esenzione Eta Max"
          },
          "esenzione_notti_oltre": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Esenzione Notti Oltre"
          },
          "importo_cent": {
            "title": "Importo Cent",
            "type": "integer"
          },
          "periodicita": {
            "$ref": "#/components/schemas/Periodicita"
          }
        },
        "required": [
          "importo_cent",
          "periodicita",
          "esenzione_eta_max",
          "esenzione_notti_oltre"
        ],
        "title": "ParametriTassaOutput",
        "type": "object"
      },
      "Periodicita": {
        "enum": [
          "mensile",
          "trimestrale",
          "semestrale",
          "annuale"
        ],
        "title": "Periodicita",
        "type": "string"
      },
...
      "StatoConfigurazione": {
        "enum": [
          "configurata",
          "configurazione_non_disponibile"
        ],
        "title": "StatoConfigurazione",
        "type": "string"
      },
...
      "StatoStruttura": {
        "description": "Una Struttura si archivia, mai si distrugge (AD-20).",
        "enum": [
          "attiva",
          "archiviata"
        ],
        "title": "StatoStruttura",
        "type": "string"
      },
      "StrutturaInput": {
        "properties": {
          "cin": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cin"
          },
          "comune": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Comune",
            "type": "string"
          },
          "comune_codice_istat": {
            "anyOf": [
              {
                "maxLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comune Codice Istat"
          },
          "nome": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Nome",
            "type": "string"
          },
          "regione": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Regione",
            "type": "string"
          }
        },
        "required": [
          "nome",
          "comune",
          "regione"
        ],
        "title": "StrutturaInput",
        "type": "object"
      },
      "StrutturaOutput": {
        "properties": {
          "cin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cin"
          },
          "cin_mancante": {
            "readOnly": true,
            "title": "Cin Mancante",
            "type": "boolean"
          },
          "comune": {
            "title": "Comune",
            "type": "string"
          },
          "comune_codice_istat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comune Codice Istat"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "regione": {
            "title": "Regione",
            "type": "string"
          },
          "regione_codice_istat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regione Codice Istat"
          },
          "stato": {
            "$ref": "#/components/schemas/StatoStruttura"
          }
        },
        "required": [
          "id",
          "nome",
          "comune",
          "regione",
          "cin",
          "stato",
          "comune_codice_istat",
          "regione_codice_istat",
          "cin_mancante"
        ],
        "title": "StrutturaOutput",
        "type": "object"
      },
      "StrutturaUpdate": {
        "properties": {
          "cin": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cin"
          },
          "comune": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comune"
          },
          "comune_codice_istat": {
            "anyOf": [
              {
                "maxLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comune Codice Istat"
          },
          "nome": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "regione": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regione"
          }
        },
        "title": "StrutturaUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Gestionale per host privati di affitti brevi: calendario unificato, adempimenti italiani, regole di prezzo, operatività.",
    "title": "HostPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/strutture": {
      "get": {
        "operationId": "lista_api_v1_strutture_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrutturaOutput"
                  },
                  "title": "Response Lista Api V1 Strutture Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Lista",
        "tags": [
          "strutture"
        ]
      },
      "post": {
        "operationId": "crea_api_v1_strutture_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrutturaInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrutturaOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Crea",
        "tags": [
          "strutture"
        ]
      }
    },
    "/api/v1/strutture/{struttura_id}": {
      "patch": {
        "operationId": "aggiorna_api_v1_strutture__struttura_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "struttura_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Struttura Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrutturaUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrutturaOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aggiorna",
        "tags": [
          "strutture"
        ]
      }
    },
    "/api/v1/strutture/{struttura_id}/archivia": {
      "post": {
        "operationId": "archivia_api_v1_strutture__struttura_id__archivia_post",
        "parameters": [
          {
            "in": "path",
            "name": "struttura_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Struttura Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrutturaOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archivia",
        "tags": [
          "strutture"
        ]
      }
    },
    "/api/v1/strutture/{struttura_id}/configurazione-normativa": {
      "get": {
        "description": "Configurazione applicabile alla Struttura, con degrado sicuro (AD-9).",
        "operationId": "configurazione_normativa_api_v1_strutture__struttura_id__configurazione_normativa_get",
        "parameters": [
          {
            "in": "path",
            "name": "struttura_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Struttura Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "alla_data",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Alla Data"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurazioneNormativaOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configurazione Normativa",
        "tags": [
          "strutture"
        ]
      }
    }
  }
}